+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 - Random space substitution transformer
 - Random case swapping transformer
 - Random comment rewriting transformer
+- Random comment removing transformer
 - Random fuzzer